backend/requirements.txt
flask
flask-cors
PyPDF2
pdfplumber
google-generativeai
python-dotenv